+++ OpenRA.Mods.Cnc/Widgets/Logic/ProductionTabsLogic.cs
@@ -9,7 +9,7 @@
 #endregion
 
 using System;
-using OpenRA.Mods.RA.Widgets;
+using OpenRA.Mods.Common.Widgets;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Cnc.Widgets.Logic
