+++ OpenRA.Mods.Common/Traits/Render/RenderBuildingWarFactory.cs
@@ -12,10 +12,9 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Graphics;
-using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA.Traits
+namespace OpenRA.Mods.Common.Traits
 {
 	class RenderBuildingWarFactoryInfo : RenderBuildingInfo
 	{
