+++ OpenRA.Mods.Common/Widgets/Logic/ProductionTooltipLogic.cs
@@ -15,7 +15,7 @@ using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 using OpenRA.Widgets;
 
-namespace OpenRA.Mods.RA.Widgets.Logic
+namespace OpenRA.Mods.Common.Widgets.Logic
 {
 	public class ProductionTooltipLogic
 	{
