+++ OpenRA.Mods.Common/Widgets/ProductionPaletteWidget.cs
@@ -14,12 +14,11 @@ using System.Drawing;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Orders;
-using OpenRA.Mods.RA.Traits;
+using OpenRA.Mods.Common.Orders;
 using OpenRA.Network;
 using OpenRA.Widgets;
 
-namespace OpenRA.Mods.RA.Widgets
+namespace OpenRA.Mods.Common.Widgets
 {
 	public class ProductionIcon
 	{
