+++ OpenRA.Mods.D2k/Traits/World/ChooseBuildTabOnSelect.cs
@@ -9,8 +9,8 @@
 #endregion
 
 using System.Linq;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.D2k.Widgets;
-using OpenRA.Mods.RA;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 using OpenRA.Widgets;
