+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -135,16 +135,10 @@
     <Compile Include="Orders\PlaceBuildingOrderGenerator.cs" />
     <Compile Include="Orders\GlobalButtonOrderGenerator.cs" />
     <Compile Include="ParaDrop.cs" />
-    <Compile Include="Player\ClassicProductionQueue.cs" />
-    <Compile Include="Player\PlaceBuilding.cs" />
-    <Compile Include="Player\ProductionQueue.cs" />
     <Compile Include="Traits\PortableChrono.cs" />
     <Compile Include="Scripting\Properties\GuardProperties.cs" />
     <Compile Include="Widgets\Logic\TabCompletionLogic.cs" />
-    <Compile Include="Production.cs" />
-    <Compile Include="ProductionBar.cs" />
     <Compile Include="Traits\Render\RenderJammerCircle.cs" />
-    <Compile Include="Render\RenderBuildingWarFactory.cs" />
     <Compile Include="Render\RenderHarvester.cs" />
     <Compile Include="Render\RenderDisguise.cs" />
     <Compile Include="Render\RenderLandingCraft.cs" />
@@ -161,7 +155,6 @@
     <Compile Include="TraitsInterfaces.cs" />
     <Compile Include="Traits\Buildings\Fake.cs" />
     <Compile Include="Traits\Buildings\OreRefinery.cs" />
-    <Compile Include="Traits\Buildings\PrimaryBuilding.cs" />
     <Compile Include="Traits\Harvester.cs" />
     <Compile Include="Traits\HarvesterHuskModifier.cs" />
     <Compile Include="Traits\LeavesHusk.cs" />
@@ -244,10 +237,6 @@
     <Compile Include="SupportPowers\SpawnActorPower.cs" />
     <Compile Include="Traits\Buildings\ClonesProducedUnits.cs" />
     <Compile Include="Cloneable.cs" />
-    <Compile Include="Widgets\ProductionPaletteWidget.cs" />
-    <Compile Include="Widgets\ProductionTabsWidget.cs" />
-    <Compile Include="Widgets\ProductionTypeButtonWidget.cs" />
-    <Compile Include="Widgets\Logic\ProductionTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\Ingame\IngameCashCounterLogic.cs" />
     <Compile Include="Widgets\Logic\Ingame\IngamePowerCounterLogic.cs" />
     <Compile Include="Widgets\Logic\Ingame\IngamePowerBarLogic.cs" />
@@ -258,7 +247,6 @@
     <Compile Include="Widgets\Logic\SupportPowerTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\SupportPowerBinLogic.cs" />
     <Compile Include="Widgets\Logic\DebugMenuLogic.cs" />
-    <Compile Include="ProductionQueueFromSelection.cs" />
     <Compile Include="GainsStatUpgrades.cs" />
     <Compile Include="Warheads\PerCellDamageWarhead.cs" />
     <Compile Include="Scripting\Properties\HarvesterProperties.cs" />
